Major aestetics update to the mailing system.
All the emails from Inforequest have been updated to the look'n'feel of
the site. Some cofiguration have been added to the config_inc.php, and
the custom_functions_inc.php has all the heavy load of the html code
used on the the email structure to look clean and modern.
2 bugs coorected where the internationalization was not being used for
the emails in bug_update and bug_api.
Some documentation on this changes have been added.