Empty user list when "Hide inactive" option is checked
The refactoring of db_helper_compare_days() function in commit 0a33bdf
introduced usage of a new query
parameter; the parameters list is not built in the order expected by the
generated SQL query, causing the boolean value for "enabled" to be used
as a date, and the given timestamp is compared with "enabled" flag.

The SQL now applies the "disabled condition" where clause before the
date filter, so the order of parameters matches.

Also removed useless db_now() parameter assignment at line 249, the
target date parameter has already been added at line 219.

And crunched lines 123 and 125 for compatibility issues with master
branch.

Fixes #0000049
Ref. (master-Jun 2, 2015)
85f94206a648d1651c5d23e7bca83b0bd6250643