Restrict user list on the bug_change_status_page
Restrict user list on the bug_change_status_page.php acording to
workflow permisions over user profiles.

A new function was created on access_api.php to handle this, and the new
configuration option on config_defaults_inc.php was included ( as
x[array [ array ] ] ) with the values of the profiles of the users that
must be shown on thje moment of a status change acording to the status
issued.

Fixes #54