Fix SQL error when custom field is in filter/sort
If the same custom field is used for filtering to a value that is not
"any" and is also used for sorting,
the generated query defines the same table alias twice causing a SQL
error.
The fix is to use different aliasing prefix for filtering vs. sorting.

Fixes #76
Ref. (master-Abr 27, 2015)
53ee0c749e38d2b7afd5e38095c475004ec04187