Fix system notice in installer
When the installer is not able to connect to the database as admin, the
DB version check fails because the version cannot be retrieved from the
server. This causes display of a PHP notice. Initializing
$t_version_info fixes this problem.

Additionally, the test result is shown as 'BAD' in this case, which is
not necessarily correct (we just don't know), so we now display
'POSSIBLE PROBLEM' instead.

Also multi_sort: use closure instead of custom_function()
Ref: master (10-Apr-2015)
c4bf4a09e168181cdd4d1d973e736328e00ca1d6

Fixes #83

Ref: master (29-Abr-2015)
5fea91cc88e9a31303f96438c7b11f90ab4cda1c