Added suport for a second ms-word report
Added suport for a second word report and export acording to
organization especs to match the requeriments of the whitepapers.

The word report was duplicated and the option was added to the
print_all_bug_page named word2, strings was added only to english and
spanish.

TO-DO: This change needs to be improved with a config option to select
the columns that need to be in this ms-word export.