Change to table Estadísticas por Desarrollador in
Summary page.
This changes makes it posible for this summary to include the issues
that developers worked on but they didn't closed, or simply helped in
some way. Making this statistic real according to what work was actualy
done by the developer.

New helper_api.php function to support this changes.