Changes to bugnotes and new DevLog
New Developer log created to allow them to documentate changes made on
the requeriments.

Many files changed and a new table on the DB created, the schema is on
comentaries but it signals the structure of the table.

Also the privacy for the Bugnotes has now 3 levels acording to
configuration.